Block other threads from checking a URL if it's already being checked
Multiple threads can be queued up to validate URLs. So two threads can
see an empty cache and then both run validation.

This PR sets the first thread to put a sentinel value into the cache
that the other threads will spin on, so we actually achieve the result
of only every trying to validate a given URL one time.